Use the same index for MML subroutines and compiled subroutines
This will allow the subroutine call graph to be built from Commands.

+++ crates/compiler/src/command_compiler/commands.rs
@@ -263,7 +263,7 @@ pub(crate) enum Command {
 
     DisableNoise,
 
-    CallSubroutine(usize, SubroutineCallType),
+    CallSubroutine(u8, SubroutineCallType),
     StartLoop,
     SkipLastLoop,
     EndLoop(LoopCount),
